+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add option to disable safety settings
 - Add option to exclude locales
 - Fix quota exceeded error handling for Vertex AI
+- Fix setting use escaping and relax syntax in `l10n.yaml`
 
 ## 0.1.2
 - Fix setting model provider in `l10n.yaml`